# src/openapi/openapi.gen.yaml
openapi: 3.0.0
info:
  title: veryExpress generated api server
  description: OpenApi Spec generated by veryExpress
  version: 1.0.0
paths:
...
  /api/user:
    x-documentName: User
    post:
      operationId: postUser
      tags:
        - user
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters: []
      requestBody:
        description: post User
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/postUserBody'
      responses:
        '201':
          description: Crerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/postUserResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
  /api/user/search:
    x-documentName: User
    post:
      operationId: getListUser
      tags:
        - user
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters: []
      requestBody:
        description: getList User
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getListUserBody'
      responses:
        '201':
          description: Crerated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getListUserResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
  /api/user/{id}:
    x-documentName: User
    get:
      operationId: getUser
      tags:
        - user
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            x-format: ObjectId
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getUserResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
    patch:
      operationId: patchUser
      tags:
        - user
      security:
        - BearerAuth: []
        - AuthIndex: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            x-format: ObjectId
      requestBody:
        description: patch User
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/patchUserBody'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/patchUserResponse'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '405':
          description: Method Not Allowed
        '413':
          description: Payload Too Large
        '429':
          description: Too Many Requests
        '500':
          description: Internal Server Error
        '502':
          description: Bad Gateway
        '503':
          description: Service Unavailable
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AuthIndex:
      type: apiKey
      in: header
      name: X-Auth-Index
  schemas:
...
    getUserResponse:
      type: object
      properties:
        _id:
          type: string
        authProfiles:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              authId:
                type: string
              username:
                type: string
              password:
                type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
        roles:
          type: array
          items:
            type: string
            enum:
              - user
          default:
            - user
        profileErrors:
          type: array
          items:
            type: string
    getListUserBody:
      type: object
      properties:
        _filter:
          type: object
        _select:
          type: array
          items:
            type: string
        _join:
          type: array
          items:
            type: string
      example:
        _filter:
          fieldName:
            $regex: str
            $options: i
        _select:
          - FieldName
        _join:
          - CollectionName
    getListUserResponse:
      type: array
      items:
        type: object
        properties:
          _id:
            type: string
          authProfiles:
            type: array
            items:
              type: object
              properties:
                provider:
                  type: string
                authId:
                  type: string
                username:
                  type: string
                password:
                  type: string
          email:
            type: string
            format: email
          name:
            type: string
            maxLength: 64
          active:
            type: boolean
            default: true
          locale:
            type: string
            maxLength: 8
          roles:
            type: array
            items:
              type: string
              enum:
                - user
            default:
              - user
          profileErrors:
            type: array
            items:
              type: string
    postUserBody:
      type: object
      properties:
        authProfiles:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              authId:
                type: string
              username:
                type: string
              password:
                type: string
            required: []
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
        roles:
          type: array
          items:
            type: string
            enum:
              - user
          default:
            - user
        profileErrors:
          type: array
          items:
            type: string
      required:
        - name
        - active
    postUserResponse:
      type: object
      properties:
        _id:
          type: string
        authProfiles:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              authId:
                type: string
              username:
                type: string
              password:
                type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
        roles:
          type: array
          items:
            type: string
            enum:
              - user
          default:
            - user
        profileErrors:
          type: array
          items:
            type: string
    patchUserBody:
      type: object
      properties:
        authProfiles:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              authId:
                type: string
              username:
                type: string
              password:
                type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
        roles:
          type: array
          items:
            type: string
            enum:
              - user
          default:
            - user
        profileErrors:
          type: array
          items:
            type: string
    patchUserResponse:
      type: object
      properties:
        _id:
          type: string
        authProfiles:
          type: array
          items:
            type: object
            properties:
              provider:
                type: string
              authId:
                type: string
              username:
                type: string
              password:
                type: string
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 64
        active:
          type: boolean
          default: true
        locale:
          type: string
          maxLength: 8
        roles:
          type: array
          items:
            type: string
            enum:
              - user
          default:
            - user
        profileErrors:
          type: array
          items:
            type: string